+++ cubiomes/CHANGELOG.md
@@ -6,11 +6,13 @@ The project follows semver
 
 ## [Unreleased]
 
+## [0.3.1] - 2025-03-05
+
 ### Added
 - Heightamp generation
 
 ### Fixed
-- Invalid test (gh issues #9, #10, #11)
+- Invalid tests (gh #9, #10, #11)
 
 ## [0.3.0] - 2025-03-02